Add files via upload
This folder contains all the necessary assets for the Flappy Bird game created using Python and Pygame. These assets should be downloaded and placed in the same directory as the main game program for everything to work correctly.

Assets Include:

Bird Sprites: Images of the bird in various flap positions for smooth animation.

Pipe Images: Images of the top and bottom pipes to serve as obstacles in the game.

Background Image: A background image to set the scene for the game.

Ground Image: An image for the ground that scrolls beneath the bird during gameplay.

Audio Files:

Flap sound for when the bird jumps.

Collision sound when the bird hits an obstacle or the ground.

Point sound when the bird successfully passes through a set of pipes.

Make sure to download and place all these files into the same folder as the game script to ensure proper game functionality.