Add files via upload
Flappy Bird Game Executable (.exe)
This is the compiled executable version of the Flappy Bird game developed using Python and Pygame. The .exe file allows you to run the game on a Windows system without the need for Python or Pygame installation.

Instructions:

Download the .exe file along with the required assets (images and audio).
Ensure that the assets are placed in the same folder as the .exe file.
Run the .exe to start playing the game.
This provides an easy and portable way to share and play the game without requiring the source code setup.